refactor(admin): decompose audio page (738→187) onto shared primitives, drop dead cleanup UI
Table moves to features/admin/audio/components (audio-table,
audio-stats-cards; delete-dialog now wraps shared ConfirmDialog). The
four TODO-marked dead elements — always-zero Unused/Cleanup stat cards,
the '-'-only Usage column, and the fake 'Coming Soon' cleanup dialog —
are deleted.
